+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Always reference the ticket number at the end of the issue description.
 
+##[unreleased]
+
+## Changed
+- added pagination_legend block and show_legend pagination option - [#277][277]
+
+[277]: //github.com/sanoma/django-arctic/issues/277
+
+
 ## 1.0.1 (2017-03-01)
 
 ### Fixed