+++ CHANGELOG.md
@@ -8,12 +8,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 Always reference the ticket number at the end of the issue description.
 
 
-##[unreleased]
+## [unreleased]
+
+### Fixed
+
+- Search bar items now join like a grouped button - [#267][267]
 
 ## Changed
+
 - added pagination_legend block and show_legend pagination option - [#277][277]
 - added virtual_ordering_fields for ListView - [#274][274]
 
+[267]: //github.com/sanoma/django-arctic/issues/267
 [277]: //github.com/sanoma/django-arctic/issues/277
 [274]: //github.com/sanoma/django-arctic/issues/274